# Modify execute_action to stream results